Increased accuracy and figured out the timescale (us), then converted it to microseconds

+++ src/main/java/com/squarepolka/digiscope/plot/PlotPointDigital.java
@@ -1,16 +1,18 @@
 package com.squarepolka.digiscope.plot;
 
+import java.math.BigDecimal;
+
 public class PlotPointDigital {
 
-    private double timeValue;
+    private BigDecimal timeValue;
     private boolean oneOrZero;
 
-    public PlotPointDigital(double timeValue, boolean oneOrZero) {
+    public PlotPointDigital(BigDecimal timeValue, boolean oneOrZero) {
         this.timeValue = timeValue;
         this.oneOrZero = oneOrZero;
     }
 
-    public double getTimeValue() {
+    public BigDecimal getTimeValue() {
         return timeValue;
     }
